Refactoring.  Added bookID for check of gateway language list.
